Send PayJoin from LND wallet
This entails creating an Original PSBT and PayJoin Proposal PSBT
after negotiating with the PayJoin Uri. The integration test is
modified to send_payjoin using the Scheduler rather than accessing the
bip78 payjoin library directly.

In order to test locally, the payjoin https client has to be configured
to accept a self-signed certificate ONLY while the test runs.